# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ocolis",
    "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. In quis aliquam neque. Maecenas non pretium quam. Etiam sapien metus, bibendum eget neque nec, luctus maximus lacus. Vestibulum bibendum ipsum at ante mollis, id tristique massa eleifend. Etiam eu ipsum eros. Fusce ac fringilla ligula. Aenean fermentum augue nisl, eu congue tellus fermentum ut. Sed sodales, est vel sagittis faucibus, lorem velit rutrum tellus, vitae sagittis ex erat nec odio. Nam consequat libero in ante feugiat, nec ornare nunc hendrerit.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:7070"
    }
  ],
...
  "paths": {
...
    "/package": {
      "get": {
        "tags": [
          "colis"
        ],
        "summary": "Récupère tous les colis stockés en base de données",
        "operationId": "getAllPackages",
        "responses": {
          "200": {
            "description": "Succès de la requête",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Package"
                }
              }
            }
          },
          "404": {
            "description": "Aucun colis stocké en base de données"
          }
        }
      },
      "post": {
        "tags": [
          "colis"
        ],
        "summary": "Création d'un nouveau colis",
        "operationId": "createPackage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Informations concernant le nouveau colis. Tous les champs sont requis. \"weight\", \"volume\", \"worth\", \"sender_id\", \"recipient_id\" et \"expedition_id\" doivent être des entiers et les identifiants correspondants à \"sender_id\", \"recipient_id\" ou \"expedition_id\" doivent correspondre à une entrée en base de données",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "serial_number": {
                    "type": "string",
                    "format": "byte"
                  },
                  "content_description": {
                    "type": "string",
                    "format": "byte"
                  },
                  "weight": {
                    "type": "number",
                    "format": "float"
                  },
                  "volume": {
                    "type": "number",
                    "format": "int64"
                  },
                  "worth": {
                    "type": "number",
                    "format": "int64"
                  },
                  "sender_id": {
                    "type": "number",
                    "format": "int64"
                  },
                  "recipient_id": {
                    "type": "number",
                    "format": "int64"
                  },
                  "expedition_id": {
                    "type": "number",
                    "format": "int64"
                  }
                },
                "required": [
                  "serial_number",
                  "content_description",
                  "weight",
                  "volume",
                  "worth",
                  "sender_id",
                  "recipient_id",
                  "expedition_id"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Colis crée. Renvoie l'id du nouveau collis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "example": 4
                }
              }
            }
          },
          "400": {
            "description": "Echec de la requête. Un des champs est manquant, \"weight\", \"volume\", \"worth\", \"sender_id\", \"recipient_id\" ou \"expedition_id\" n'est pas un entier ou l'identifiant correspondant à \"sender_id\", \"recipient_id\" ou \"expedition_id\" ne correspond à aucune entrée en base de données"
          },
          "401": {
            "description": "Le client n'est pas authentifié, le token n'est pas valide/expiré ou l'utilisateur n'existe plus"
          }
        }
      }
    },
    "/package/{packageId}": {
      "get": {
        "tags": [
          "colis"
        ],
        "summary": "Récupère un colis selon son id",
        "operationId": "getPackageById",
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "description": "Id du colis recherché",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de la requête",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Package"
                }
              }
            }
          },
          "404": {
            "description": "Colis inexistant/Id incorrect"
          }
        }
      },
      "patch": {
        "tags": [
          "colis"
        ],
        "summary": "Modifie un colis selon son id",
        "operationId": "updatePackageById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "description": "Id du colis à modifier",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Nouvelles informations pour le colis indiqué. \"weight\", \"volume\", \"worth\", \"sender_id\", \"recipient_id\" et \"expedition_id\" doivent être des entiers et les identifiants correspondants à \"sender_id\", \"recipient_id\" ou \"expedition_id\" doivent correspondre à une entrée en base de données",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "serial_number": {
                    "type": "string",
                    "format": "byte"
                  },
                  "content_description": {
                    "type": "string",
                    "format": "byte"
                  },
                  "weight": {
                    "type": "number",
                    "format": "float"
                  },
                  "volume": {
                    "type": "number",
                    "format": "int64"
                  },
                  "worth": {
                    "type": "number",
                    "format": "int64"
                  },
                  "sender_id": {
                    "type": "number",
                    "format": "int64"
                  },
                  "recipient_id": {
                    "type": "number",
                    "format": "int64"
                  },
                  "expedition_id": {
                    "type": "number",
                    "format": "int64"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès de la requête. L'attribut \"modified\" permet d'indiquer si le changement a été effectif ou non",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "modified": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Echec de la requête. \"weight\", \"volume\", \"worth\", \"sender_id\", \"recipient_id\" ou \"expedition_id\" n'est pas un entier ou l'identifiant correspondant à \"sender_id\", \"recipient_id\" ou \"expedition_id\" ne correspond à aucune entrée en base de données"
          },
          "401": {
            "description": "Le client n'est pas authentifié, le token n'est pas valide/expiré ou l'utilisateur n'existe plus"
          },
          "404": {
            "description": "Colis inexistant/Id incorrect"
          }
        }
      },
      "delete": {
        "tags": [
          "colis"
        ],
        "summary": "Supprime un colis selon son id",
        "operationId": "deletePackageById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "description": "Id du colis à supprimer",
            "required": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès de la requête. L'attribut \"deleted\" permet d'indiquer si le changement a été effectif ou non",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          },
          "401": {
            "description": "Le client n'est pas authentifié, le token n'est pas valide/expiré ou l'utilisateur n'existe plus"
          },
          "404": {
            "description": "Colis inexistant/Id incorrect"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Place": {
        "required": [
          "id",
          "reference",
          "name",
          "address",
          "postal_code",
          "city"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "reference": {
            "type": "string",
            "format": "byte"
          },
          "name": {
            "type": "string",
            "format": "byte"
          },
          "address": {
            "type": "string",
            "format": "byte"
          },
          "additional": {
            "type": "string",
            "format": "byte"
          },
          "postal_code": {
            "type": "integer",
            "format": "int64"
          },
          "city": {
            "type": "string",
            "format": "byte"
          }
        },
        "xml": {
          "name": "Place"
        }
      },
      "Expedition": {
        "required": [
          "id",
          "driver_name",
          "vehicle_plate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "driver_name": {
            "type": "string",
            "format": "byte"
          },
          "vehicle_plate": {
            "type": "string",
            "format": "byte"
          },
          "starting_time": {
            "type": "string",
            "format": "date-time"
          },
          "ending_time": {
            "type": "string",
            "format": "date-time"
          }
        },
        "xml": {
          "name": "Expedition"
        }
      },
      "Package": {
        "required": [
          "id",
          "serial_number",
          "content_description",
          "weight",
          "volume",
          "worth",
          "sender_id",
          "recipient_id",
          "expedition_id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "serial_number": {
            "type": "string",
            "format": "byte"
          },
          "content_description": {
            "type": "string",
            "format": "byte"
          },
          "weight": {
            "type": "integer",
            "format": "float"
          },
          "volume": {
            "type": "integer",
            "format": "int64"
          },
          "worth": {
            "type": "integer",
            "format": "int64"
          },
          "sender_id": {
            "$ref": "#/components/schemas/Place"
          },
          "recipient_id": {
            "$ref": "#/components/schemas/Place"
          },
          "expedition_id": {
            "$ref": "#/components/schemas/Expedition"
          }
        },
        "xml": {
          "name": "Package"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}